pyproject.toml
[project]
name = "agentic-ai-immersion"
version = "0.1.0"
description = "Agentic AI Immersion Workshop — lean setup for hands-on labs"
requires-python = ">=3.11,<3.14"

dependencies = [
    "azure-ai-projects>=2.0.0b1",
    "azure-identity>=1.15.0",
    "azure-search-documents>=11.7.0b2",
    "python-dotenv>=1.0.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
list-agents = "labs.list_agents:main"
list-vector-stores = "labs.list_vector_stores:main"
agents-lab01 = "labs.agents.lab01_basics:main"
agents-lab02 = "labs.agents.lab02_code_interpreter:main"
agents-lab03 = "labs.agents.lab03_file_search:main"
obs-lab01 = "labs.observability.lab01_telemetry:main"
obs-lab02 = "labs.observability.lab02_agent_evaluation:main"
obs-lab03 = "labs.observability.lab03_function_tools_eval:main"
hosted-lab01 = "labs.hosted.lab01_hosted_agents:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true
prerelease = "if-necessary-or-explicit"
